% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertColumnTypes.R
\name{convertColumnTypes}
\alias{convertColumnTypes}
\title{Titel}
\usage{
convertColumnTypes(inputDF, searchType, targetType)
}
\arguments{
\item{inputDF}{A data.frame}

\item{searchType}{Type of the columns which have to be converted.}

\item{targetType}{Type to which the columns with type "searchType" will be converted}
}
\value{
inputDF as a data.table
}
\description{
"Converts all columns of a data frame with type X to type Y."
}
\details{
Untertitel
}
\examples{
\dontrun{dt=dfTodt(as.data.frame(matrix(1:100, ncol=10)))}
}
